Se creó el fragment para el consumo del servicio api

+++ app/src/main/java/com/jonathan/empowerment/domain/model/Us.kt
@@ -3,7 +3,7 @@ package com.jonathan.empowerment.domain.model
 import com.jonathan.empowerment.data.datasource.local.entity.UsEntity
 
 data class Us(
-    val amount: Long?,
+    val amount: Float?,
     val unitShort: String?,
     val unitLong: String?
 )
